refactor(parser): split algebra/expressions/types into parser_pkg/ submodules
Phase 1B Batch 7 (Moves 15-17 from docs/PHASE1-PLAN.md).  Extract
the relational-algebra rules, the full expression grammar, and the
type-expression / generic-params rules from parser.py into three
new mixin modules under src/txt2tex/parser_pkg/.

* algebra.py — _parse_restrict (sigma), _parse_project (pi),
  _parse_relation_rename (R[B/A]), the GROUP family (_group_rhs,
  _group_regroup_rhs, _group_aggregate_rhs, _aggregator_clause,
  _ungroup_rhs), and the _is_attr_name_token helper.
* expressions.py — the full recursive-descent expression grammar
  from _parse_expr down to _parse_atom: conditionals, iff/implies,
  or/and, unary, additive, multiplicative, range, comparison,
  relation / set ops, union, cross (and natural-join branch),
  intersect, postfix, plus quantifier / lambda / binding /
  set-comprehension parsers and the atom literals (sequence, bag).
* types.py — _parse_function_type, _parse_generic_instantiation,
  _parse_generic_params.

Gate 7 verified:
* make check — clean.
* make test-e2e — 159/159 byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

parser.py shrinks from 4666 → 1467 lines (cumulative 76.9%).
