+++ src/txt2tex/parser_pkg/__init__.py
@@ -3,7 +3,10 @@
 from __future__ import annotations
 
 from txt2tex.parser_pkg import (
+    algebra,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    expressions,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    types,  # noqa: F401  # pyright: ignore[reportUnusedImport]
 )
